SVM Classifier

+++ DS/Dataset.hs
@@ -1,4 +1,4 @@
 module DS.Dataset where
-	type X = [[Int]]
-	type Y = [Int]
+	type X = [[Float]]
+	type Y = [Float]
 	data Dataset =  Dataset (X, Y)
